[refactor] PM card widget has been exported to dedicated package
