+++ apolline-flutter/lib/widgets/pm_card.dart
@@ -1,0 +1,91 @@
+import 'package:apollineflutter/services/user_configuration_service.dart';
+import 'package:apollineflutter/utils/pm_filter.dart';
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+
+class PMCard extends StatefulWidget {
+  final UserConfigurationService ucS;
+  final PMFilter indicator;
+  final List<TextInputFormatter> formatters =
+    [TextInputFormatter.withFunction((oldValue, newValue) => newValue.text.length > 3 ? oldValue : newValue)];
+  PMCard({@required this.ucS, @required this.indicator});
+
+  @override
+  State<StatefulWidget> createState() => _PMCardState();
+}
+
+
+class _PMCardState extends State<PMCard> {
+  int warningThresholdValue = 0;
+  int dangerThresholdValue = 0;
+
+  @override
+  void initState() {
+    List<int> initialThresholdsValues = widget.ucS.userConf.getThresholds(widget.indicator);
+    warningThresholdValue = initialThresholdsValues[0];
+    dangerThresholdValue = initialThresholdsValues[1];
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+
+    return Card(
+        margin: EdgeInsets.only(bottom: 20),
+        child: Wrap(
+          children: [
+            Container(
+                padding: EdgeInsets.only(left: 15, top: 10, bottom: 20),
+                child: Text(widget.indicator.getLabelKey().tr(), style: TextStyle(
+                    fontSize: 16,
+                    fontWeight: FontWeight.bold
+                ))
+            ),
+            ListTile(
+              title: Text("Warning threshold"),
+              trailing: Container(
+                width: 80,
+                child: TextFormField(
+                  keyboardType: TextInputType.number,
+                  inputFormatters: widget.formatters,
+                  initialValue: warningThresholdValue.toString(),
+                  onFieldSubmitted: (value) {
+                    widget.ucS.userConf.updatePMThreshold(widget.indicator, 0, int.parse(value));
+                    widget.ucS.update();
+                  },
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: "15",
+                      suffixIcon: Text("µm/m³")
+                  ),
+                ),
+              ),
+            ),
+            ListTile(
+              title: Text("Danger threshold"),
+              trailing: Container(
+                width: 80,
+                child: TextFormField(
+                  keyboardType: TextInputType.number,
+                  inputFormatters: widget.formatters,
+                  initialValue: dangerThresholdValue.toString(),
+                  onFieldSubmitted: (value) {
+                    widget.ucS.userConf.updatePMThreshold(widget.indicator, 1, int.parse(value));
+                    widget.ucS.update();
+                  },
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: "30",
+                      suffixIcon: Text("µm/m³")
+                  ),
+                ),
+              ),
+            )
+          ],
+        )
+    );
+  }
+}
